+++ frb_macros/src/components/mod.rs
@@ -1,3 +1,3 @@
-pub(crate) mod stripper;
+pub(crate) mod converter;
 pub(crate) mod attr_external;
 pub(crate) mod encoder;
